It works

+++ fdf/libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: eschmitz <eschmitz@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/08 17:58:18 by eschmitz          #+#    #+#             */
-/*   Updated: 2024/07/08 18:30:29 by eschmitz         ###   ########.fr       */
+/*   Updated: 2024/07/15 15:18:22 by eschmitz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
 	size_t	len;
